conditionally validate identity password
An identity password is only required if password_use is set to 'custom'. All other values for password_use will tell the Migadu API to ignore the password. We are a bit stricter here and enforce that no password is set if password_use is not set to 'custom'.

Relates to #145

Signed-off-by: Sebastian Hoß <seb@xn--ho-hia.de>